% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finmath.R
\name{annuity_due}
\alias{annuity_due}
\title{Menghitung Annuity Due}
\usage{
annuity_due(payment, rate, n)
}
\arguments{
\item{payment}{Pembayaran yang dilakukan setiap periode.}

\item{rate}{Tingkat bunga per periode (dalam desimal).}

\item{n}{Jumlah periode pembayaran.}
}
\value{
Nilai Annuity Due, yaitu total nilai dari serangkaian pembayaran dengan bunga yang dibayarkan di setiap awal periode.
}
\description{
Fungsi ini menghitung nilai Annuity Due, yaitu serangkaian pembayaran yang dilakukan di awal setiap periode.
}
