% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finmath.R
\name{annuity_immediate}
\alias{annuity_immediate}
\title{Menghitung Annuity Immediate}
\usage{
annuity_immediate(payment, rate, n)
}
\arguments{
\item{payment}{Pembayaran per periode}

\item{rate}{Tingkat bunga per periode}

\item{n}{Jumlah periode}
}
\value{
Nilai Annuity Immediate, yaitu total nilai dari serangkaian pembayaran dengan bunga yang dibayarkan di setiap akhir periode.
}
\description{
Menghitung nilai sekarang dari anuitas yang pembayarannya dilakukan di setiap akhir periode.
}
