% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finmath.R
\name{compound_interest}
\alias{compound_interest}
\title{Menghitung Compound Interest}
\usage{
compound_interest(principal, rate, time, comp_per_year = 1)
}
\arguments{
\item{principal}{Jumlah pokok awal.}

\item{rate}{Tingkat bunga per periode (dalam desimal)}

\item{time}{Lama waktu}

\item{comp_per_year}{Frekuensi kapitalisasi bunga dalam setahun (default = 1).}
}
\value{
Sebuah list berisi amount(total nilai setelah bunga majemuk dihitung) dan interest(jumlah bunga yang diperoleh selama periode).
}
\description{
Fungsi untuk menghitung jumlah akhir dan bunga berdasarkan bunga majemuk.
}
