% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finmath.R
\name{effective_discount}
\alias{effective_discount}
\title{Menghitung Effective Discount}
\usage{
effective_discount(rate)
}
\arguments{
\item{rate}{Tingkat bunga (dalam desimal)}
}
\value{
Diskon efektif
}
\description{
Menghitung tingkat diskon efektif
}
