% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finmath.R
\name{future_value}
\alias{future_value}
\title{Menghitung Future Value}
\usage{
future_value(principal, rate, time)
}
\arguments{
\item{principal}{Jumlah pokok awal}

\item{rate}{Tingkat bunga per periode (dalam desimal)}

\item{time}{Lama waktu}
}
\value{
Nilai masa depan
}
\description{
Menghitung nilai masa depan berdasarkan pokok awal, bunga, dan waktu.
}
