% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finmath.R
\name{loan_payment}
\alias{loan_payment}
\title{Menghitung Loan Payment}
\usage{
loan_payment(principal, rate, n)
}
\arguments{
\item{principal}{Jumlah pokok pinjaman}

\item{rate}{Tingkat bunga per periode (dalam desimal)}

\item{n}{Jumlah periode}
}
\value{
Pembayaran pinjaman per periode
}
\description{
Menghitung pembayaran pinjaman secara berkala.
}
