% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finmath.R
\name{perpetuity}
\alias{perpetuity}
\title{Menghitung Perpetuity}
\usage{
perpetuity(payment, rate)
}
\arguments{
\item{payment}{Pembayaran per periode}

\item{rate}{Tingkat bunga per periode}
}
\value{
Nilai perpetuitas
}
\description{
Menghitung nilai perpetuitas
}
