% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finmath.R
\name{present_value}
\alias{present_value}
\title{Menghitung Present Value}
\usage{
present_value(future_value, rate, time)
}
\arguments{
\item{future_value}{Nilai masa depan (future value)}

\item{rate}{Tingkat bunga per periode (dalam desimal)}

\item{time}{Lama waktu}
}
\value{
Nilai sekarang
}
\description{
Menghitung nilai sekarang berdasarkan nilai masa depan, bunga, dan waktu.
}
