% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finmath.R
\name{remaining_balance}
\alias{remaining_balance}
\title{Menghitung Remaining Balance}
\usage{
remaining_balance(principal, rate, n, t)
}
\arguments{
\item{principal}{Jumlah pokok pinjaman.}

\item{rate}{Tingkat bunga per periode (dalam desimal).}

\item{n}{Jumlah total periode.}

\item{t}{Jumlah periode yang telah dilalui.}
}
\value{
Sisa saldo pinjaman.
}
\description{
Menghitung sisa saldo pinjaman setelah beberapa pembayaran.
}
