% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finmath.R
\name{sinking_fund}
\alias{sinking_fund}
\title{Menghitung Sinking Fund}
\usage{
sinking_fund(FV, rate, m, n)
}
\arguments{
\item{FV}{Nilai masa depan (future value) yang diinginkan.}

\item{rate}{Tingkat bunga per periode (dalam desimal).}

\item{m}{Frekuensi pembayaran dalam setahun.}

\item{n}{Jumlah tahun.}
}
\value{
Pembayaran berkala untuk sinking fund.
}
\description{
Menghitung pembayaran berkala untuk sinking fund.
}
